refactor: tests to follow new conventions
