--- karat-interview/Q1/test.js
@@ -1,39 +1,0 @@
-'use strict';
-
-const { find_rectangle } = require('./submission');
-
-const image1 = [
-  [1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 0, 0, 0, 1],
-  [1, 1, 1, 0, 0, 0, 1],
-  [1, 1, 1, 1, 1, 1, 1],
-];
-
-const image2 = [
-  [1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 1, 0],
-];
-
-const image3 = [
-  [1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 0, 0],
-  [1, 1, 1, 1, 1, 0, 0],
-];
-
-test('find_rectangle - test 1', function () {
-  expect(find_rectangle(image1)).toBe([3, 5]);
-});
-
-test('find_rectangle - test 1', function () {
-  expect(find_rectangle(image2)).toBe([3, 5]);
-});
-
-test('find_rectangle - test 1', function () {
-  expect(find_rectangle(image3)).toBe([3, 5]);
-});
