+++ karat-interview/Q2/karat-q2.test.js
