+++ leetcode/Q202/test.js
@@ -2,10 +2,12 @@
 
 const { isHappy } = require('./submission');
 
-test('isHappy - test 1', function () {
-  expect(isHappy(45)).toBe(false);
-});
+describe('leetocde #202 - basic unit tests', () => {
+  test('45 is not a happy number', function () {
+    expect(isHappy(45)).toBe(false);
+  });
 
-test('isHappy - test 1', function () {
-  expect(isHappy(19)).toBe(false);
+  test('19 is not a happy number', function () {
+    expect(isHappy(19)).toBe(false);
+  });
 });
